# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  fps: number;
  canvasWidth: number;
  canvasHeight: number;
  frameCount: number;
  /** @nullable */
  thumbnailData?: string | null;
  backgroundColor?: string;
  createdAt: string;
  updatedAt: string;
}

export interface Frame {
  id: number;
  projectId: number;
  frameIndex: number;
  duration: number;
  isHold: boolean;
  /** @nullable */
  canvasData?: string | null;
  /** @nullable */
  thumbnailData?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Layer {
  id: number;
  projectId: number;
  /** @nullable */
  frameId?: number | null;
  name: string;
  layerIndex: number;
  isVisible: boolean;
  isLocked: boolean;
  opacity: number;
  blendMode?: string;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectDetail {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  fps: number;
  canvasWidth: number;
  canvasHeight: number;
  frameCount: number;
  /** @nullable */
  thumbnailData?: string | null;
  backgroundColor?: string;
  frames: Frame[];
  layers: Layer[];
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  fps?: number;
  canvasWidth?: number;
  canvasHeight?: number;
  backgroundColor?: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  fps?: number;
  backgroundColor?: string;
}

export interface ThumbnailUpdate {
  thumbnailData: string;
}

export interface FrameInput {
  frameIndex?: number;
  duration?: number;
  isHold?: boolean;
  canvasData?: string;
}

export interface FrameUpdate {
  duration?: number;
  isHold?: boolean;
  canvasData?: string;
  thumbnailData?: string;
}

export interface FrameReorder {
  frameIds: number[];
}

export interface LayerInput {
  name: string;
  frameId?: number;
  layerIndex?: number;
  opacity?: number;
  blendMode?: string;
}

export interface LayerUpdate {
  name?: string;
  isVisible?: boolean;
  isLocked?: boolean;
  opacity?: number;
  blendMode?: string;
}

export interface LayerReorder {
  layerIds: number[];
}

export interface Export {
  id: number;
  projectId: number;
  format: string;
  status: string;
  quality: string;
  fps?: number;
  /** @nullable */
  width?: number | null;
  /** @nullable */
  height?: number | null;
  /** @nullable */
  progress?: number | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileSize?: number | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ExportInputFormat =
  (typeof ExportInputFormat)[keyof typeof ExportInputFormat];

export const ExportInputFormat = {
  gif: "gif",
  mp4: "mp4",
  png_sequence: "png_sequence",
  webm: "webm",
} as const;

export type ExportInputQuality =
  (typeof ExportInputQuality)[keyof typeof ExportInputQuality];

export const ExportInputQuality = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ExportInput {
  format: ExportInputFormat;
  quality?: ExportInputQuality;
  fps?: number;
  width?: number;
  height?: number;
  transparentBackground?: boolean;
}

export interface AudioTrack {
  id: number;
  projectId: number;
  name: string;
  /** @nullable */
  audioData?: string | null;
  startFrame: number;
  volume: number;
  isMuted: boolean;
  createdAt: string;
  updatedAt: string;
}

export interface AudioTrackInput {
  name: string;
  audioData?: string;
  startFrame?: number;
  volume?: number;
}

export interface AudioTrackUpdate {
  name?: string;
  startFrame?: number;
  volume?: number;
  isMuted?: boolean;
}

export interface ActivityItem {
  type: string;
  projectName: string;
  timestamp: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalFrames: number;
  totalExports: number;
  recentActivity: ActivityItem[];
}

export type ListProjectsParams = {
  limit?: number;
  offset?: number;
};

export type GetRecentProjectsParams = {
  limit?: number;
};