# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FlipStudio Animation Studio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/frames:
    get:
      operationId: listFrames
      tags: [frames]
      summary: List all frames for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Frames list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Frame"
    post:
      operationId: createFrame
      tags: [frames]
      summary: Add a frame to a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameInput"
      responses:
        "201":
          description: Created frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frame"
  /projects/{projectId}/frames/reorder:
    patch:
      operationId: reorderFrames
      tags: [frames]
      summary: Reorder frames
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameReorder"
      responses:
        "200":
          description: Reordered
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Frame"
  /projects/{projectId}/frames/{frameId}:
    get:
      operationId: getFrame
      tags: [frames]
      summary: Get a frame
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FrameId"
      responses:
        "200":
          description: Frame data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frame"
    patch:
      operationId: updateFrame
      tags: [frames]
      summary: Update frame data (canvas data, duration, hold)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FrameId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FrameUpdate"
      responses:
        "200":
          description: Updated frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frame"
    delete:
      operationId: deleteFrame
      tags: [frames]
      summary: Delete a frame
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FrameId"
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/frames/{frameId}/duplicate:
    post:
      operationId: duplicateFrame
      tags: [frames]
      summary: Duplicate a frame
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/FrameId"
      responses:
        "201":
          description: Duplicated frame
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Frame"
...
  /projects/{projectId}/exports:
    get:
      operationId: listExports
      tags: [exports]
      summary: List exports for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Exports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Export"
    post:
      operationId: createExport
      tags: [exports]
      summary: Start an export job
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "201":
          description: Export job started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
  /projects/{projectId}/exports/{exportId}:
    get:
      operationId: getExport
      tags: [exports]
      summary: Get export job status
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - in: path
          name: exportId
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Export job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Export"
  /projects/{projectId}/audio:
    get:
      operationId: listAudioTracks
      tags: [audio]
      summary: List audio tracks for a project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Audio tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AudioTrack"
    post:
      operationId: createAudioTrack
      tags: [audio]
      summary: Add an audio track
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioTrackInput"
      responses:
        "201":
          description: Created audio track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioTrack"
  /projects/{projectId}/audio/{audioId}:
    patch:
      operationId: updateAudioTrack
      tags: [audio]
      summary: Update an audio track
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - in: path
          name: audioId
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AudioTrackUpdate"
      responses:
        "200":
          description: Updated audio track
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioTrack"
    delete:
      operationId: deleteAudioTrack
      tags: [audio]
      summary: Delete an audio track
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - in: path
          name: audioId
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema: { type: integer }
    FrameId:
      in: path
      name: frameId
      required: true
      schema: { type: integer }
...
  schemas:
...
    Frame:
      type: object
      required: [id, projectId, frameIndex, duration, isHold, createdAt, updatedAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        frameIndex: { type: integer }
        duration: { type: integer }
        isHold: { type: boolean }
        canvasData: { type: ["string", "null"] }
        thumbnailData: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    FrameInput:
      type: object
      properties:
        frameIndex: { type: integer }
        duration: { type: integer, default: 1 }
        isHold: { type: boolean, default: false }
        canvasData: { type: string }
    FrameUpdate:
      type: object
      properties:
        duration: { type: integer }
        isHold: { type: boolean }
        canvasData: { type: string }
        thumbnailData: { type: string }
    FrameReorder:
      type: object
      required: [frameIds]
      properties:
        frameIds:
          type: array
          items: { type: integer }
...
    Export:
      type: object
      required: [id, projectId, format, status, quality, createdAt, updatedAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        format: { type: string }
        status: { type: string }
        quality: { type: string }
        fps: { type: integer }
        width: { type: ["integer", "null"] }
        height: { type: ["integer", "null"] }
        progress: { type: ["number", "null"] }
        fileUrl: { type: ["string", "null"] }
        fileSize: { type: ["integer", "null"] }
        errorMessage: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ExportInput:
      type: object
      required: [format]
      properties:
        format: { type: string, enum: [gif, mp4, png_sequence, webm] }
        quality: { type: string, enum: [low, medium, high], default: high }
        fps: { type: integer }
        width: { type: integer }
        height: { type: integer }
        transparentBackground: { type: boolean, default: false }
    AudioTrack:
      type: object
      required: [id, projectId, name, startFrame, volume, isMuted, createdAt, updatedAt]
      properties:
        id: { type: integer }
        projectId: { type: integer }
        name: { type: string }
        audioData: { type: ["string", "null"] }
        startFrame: { type: integer }
        volume: { type: number }
        isMuted: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    AudioTrackInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        audioData: { type: string }
        startFrame: { type: integer, default: 0 }
        volume: { type: number, default: 1.0 }
    AudioTrackUpdate:
      type: object
      properties:
        name: { type: string }
        startFrame: { type: integer }
        volume: { type: number }
        isMuted: { type: boolean }
...