# lib/api-zod/src/generated/types/audioTrackInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface AudioTrackInput {
  name: string;
  audioData?: string;
  startFrame?: number;
  volume?: number;
}