# lib/api-zod/src/generated/types/audioTrackUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface AudioTrackUpdate {
  name?: string;
  startFrame?: number;
  volume?: number;
  isMuted?: boolean;
}