# lib/api-zod/src/generated/types/exportInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FlipStudio Animation Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ExportInputFormat } from "./exportInputFormat";
import type { ExportInputQuality } from "./exportInputQuality";

export interface ExportInput {
  format: ExportInputFormat;
  quality?: ExportInputQuality;
  fps?: number;
  width?: number;
  height?: number;
  transparentBackground?: boolean;
}